Update co_fast.py
asd

+++ co_fast.py
@@ -14,7 +14,7 @@ hopfield
 
 # Types must be given for C
 # Reads return_type(given from V_type, W_type, i_type)
-s = (int32(int32[:], double[:,:], int32))
+s = (int32(double[:], double[:,:], int32))
 
 # Update for individual units compiled to C
 @jit(s, nopython=True)
